cleanup strategise components
