+++ client/src/routes/projects/[project]/strategise/_components/MetricRow.svelte
@@ -1,0 +1,12 @@
+<script lang="ts">
+	interface Props {
+		label: string;
+		value: string;
+	}
+	let { label, value }: Props = $props();
+</script>
+
+<div class="flex justify-between text-xs">
+	<span class="text-muted-foreground">{label}:</span>
+	<span class="font-medium">{value}</span>
+</div>
