+++ client/src/routes/projects/[project]/strategise/_components/SelectedStrategyRegionCards.svelte
@@ -4,6 +4,7 @@
 	import { convertToLocaleString } from '$lib/number';
 	import type { StrategiseResult } from '$lib/types/userState';
 	import { constructRegionalMetrics } from '../utils';
+	import MetricRow from './MetricRow.svelte';
 
 	interface Props {
 		selectedStrategy: StrategiseResult;
@@ -14,12 +15,6 @@
 	let regionalMetrics = $derived(constructRegionalMetrics(selectedStrategy, populations));
 </script>
 
-{#snippet regionMetrics(label: string, value: string)}
-	<div class="flex justify-between text-xs">
-		<span class="text-muted-foreground">{label}:</span>
-		<span class="font-medium">{value}</span>
-	</div>
-{/snippet}
 <div class="mt-4">
 	<div class="grid grid-cols-2 gap-2 sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-6">
 		{#each Object.entries(regionalMetrics) as [region, { casesAverted, cost, intervention, casesAvertedPerPerson, costPerCaseAverted, costPerPerson, population }] (region)}
@@ -43,11 +38,11 @@
 						</div>
 					</div>
 					<div class="space-y-1.5 border-t border-border/30 pt-2">
-						{@render regionMetrics('Cost', `$${convertToLocaleString(cost)}`)}
-						{@render regionMetrics('Cases averted', convertToLocaleString(casesAverted))}
-						{@render regionMetrics('Cost per case averted', `$${convertToLocaleString(costPerCaseAverted)}`)}
-						{@render regionMetrics('Cost per person', `$${convertToLocaleString(costPerPerson)}`)}
-						{@render regionMetrics('Cases averted per person', convertToLocaleString(casesAvertedPerPerson))}
+						<MetricRow label="Cost" value={`$${convertToLocaleString(cost)}`} />
+						<MetricRow label="Cases averted" value={convertToLocaleString(casesAverted)} />
+						<MetricRow label="Cost per case averted" value={`$${convertToLocaleString(costPerCaseAverted)}`} />
+						<MetricRow label="Cost per person" value={`$${convertToLocaleString(costPerPerson)}`} />
+						<MetricRow label="Cases averted per person" value={convertToLocaleString(casesAvertedPerPerson)} />
 					</div>
 				</div>
 			</div>
